Got the I/O Going
Got the IO going... The clock maybe is  a bit shit. Or maybe doing the counter this way is maybe a little messed up...

To clean up the code i want to wrap all of the grid functionality in an object... Now that all this is up and running that will be a little easier to do.